requirements.txt
amqp==2.3.2
astroid==1.6.5
autopep8==1.3.5
billiard==3.5.0.3
celery==4.2.0
click==6.7
colorama==0.3.9
dj-database-url==0.5.0
dj-static==0.0.6
Django==2.0.8
django-celery-beat==1.1.1
django-celery-results==1.0.1
django-environ==0.4.4
django-extensions==2.0.7
django-heroku==0.3.1
django-richtextfield==1.2.2
django-taggit==0.22.2
gunicorn==19.8.1
isort==4.3.4
kombu==4.2.1
lazy-object-proxy==1.3.1
Markdown==2.6.11
mccabe==0.6.1
olefile==0.45.1
Pillow==5.0.0
psycopg2==2.7.5
pycodestyle==2.4.0
pylint==1.9.2
python-dotenv==0.7.1
pytz==2018.4
redis==2.10.6
six==1.11.0
static3==0.7.0
vine==1.1.4
whitenoise==3.3.1
wrapt==1.10.11
